+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [2.5.1] 30 Jan 2026
+- Fixed the campaign delivery triggers logic
+
+## [2.5.0] 29th Jan 2026
+- Fetch api and encryption support
+
+## [2.4.0] 27 Jan 2026
+- Added Nested object support in profile and event properties.
+
 ## [2.3.4] 12th Dec 2025
 - Fixed the issue where interval was not cleared in web native display.